requirements.txt
aiohttp==3.7.3
bcrypt==3.2.0
beautifulsoup4==4.9.3
bolinette==0.6.5
Jinja2==2.11.2
Markdown==3.3.3
pymongo==3.11.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
setuptools==50.3.2
SQLAlchemy==1.3.20
requests==2.25.0
